move search box to banner
