+++ frontend/src/components/Home/index.js
@@ -13,46 +13,6 @@ import {
 
 const Promise = global.Promise;
 
-class SearchBar extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      title: "",
-    };
-  }
-
-  handleSearchInput = (event) => {
-    event.preventDefault();
-    const title = event.target.value;
-    this.setState({ title });
-    if (title.length >= 3) {
-      const filteredItems = agent.Items.search(title);
-      if (filteredItems.length === 1) {
-        this.props.onSearch([]);
-      }
-      this.props.onSearch(filteredItems);
-    }
-  };
-
-  render() {
-    return (
-      <div className="search-box">
-        <span id="get-part">A place to get</span>
-        <div id="search-box">
-          <input
-            type="text"
-            className="search-input"
-            value={this.state.title}
-            onChange={this.handleSearchInput}
-            placeholder="What is that you desire?"
-          />
-        </div>
-        <span> the cool stuff.</span>
-      </div>
-    );
-  }
-}
-
 const mapStateToProps = (state) => ({
   ...state.home,
   appName: state.common.appName,
@@ -87,9 +47,8 @@ class Home extends React.Component {
   render() {
     return (
       <div className="home-page">
-        <Banner />
+        <Banner onSearch={this.props.onSearch} />
         <div className="container page">
-          <SearchBar onSearch={this.props.onSearch} />
           <Tags tags={this.props.tags} onClickTag={this.props.onClickTag} />
           <MainView />
         </div>
